fixed session for name in userpage, moved controller to DBcontroller in the controller directiv
