--- src/main/java/com/example/plantly/Domain/DBController.java
@@ -1,9 +1,0 @@
-package com.example.plantly.Domain;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class DBController {
-
-
-}
